fix(MUSD-672): fix cross-navigator navigation and deduplicate education gate
- Replace navigation.replace with navigation.navigate in education view
  since returnTo targets a screen outside the Earn stack (replace only
  works within the current navigator).
- Extract shared education-gate logic into useCashNavigation hook,
  eliminating duplication between CashSection and MusdAggregatedRow.
