+++ app/components/Views/Homepage/Sections/Cash/CashSection.tsx
@@ -6,27 +6,24 @@ import React, {
   useState,
 } from 'react';
 import { View } from 'react-native';
-import { useNavigation } from '@react-navigation/native';
 import { useSelector } from 'react-redux';
 import { Box } from '@metamask/design-system-react-native';
 import SectionHeader from '../../../../../component-library/components-temp/SectionHeader';
 import SectionRow from '../../components/SectionRow';
-import Routes from '../../../../../constants/navigation/Routes';
 import { strings } from '../../../../../../locales/i18n';
 import useHomeViewedEvent, {
   HomeSectionNames,
 } from '../../hooks/useHomeViewedEvent';
 import { useSectionPerformance } from '../../hooks/useSectionPerformance';
 import { selectIsMusdConversionFlowEnabledFlag } from '../../../../UI/Earn/selectors/featureFlags';
-import { selectMoneyHomeScreenEnabledFlag } from '../../../../UI/Money/selectors/featureFlags';
 import { useMusdConversionEligibility } from '../../../../UI/Earn/hooks/useMusdConversionEligibility';
 import { useMusdBalance } from '../../../../UI/Earn/hooks/useMusdBalance';
 import MusdAggregatedRow from './MusdAggregatedRow';
+import { useCashNavigation } from './useCashNavigation';
 
 import CashGetMusdEmptyState from './CashGetMusdEmptyState';
 import Logger from '../../../../../util/Logger';
 import { SectionRefreshHandle } from '../../types';
-import { selectMusdConversionEducationSeen } from '../../../../../reducers/user/selectors';
 
 interface CashSectionProps {
   sectionIndex: number;
@@ -42,37 +39,14 @@ const CashSection = forwardRef<SectionRefreshHandle, CashSectionProps>(
   ({ sectionIndex, totalSectionsLoaded }, ref) => {
     const sectionViewRef = useRef<View>(null);
     const [refreshVersion, setRefreshVersion] = useState(0);
-    const navigation = useNavigation();
     const isMusdConversionEnabled = useSelector(
       selectIsMusdConversionFlowEnabledFlag,
     );
     const { isEligible: isGeoEligible } = useMusdConversionEligibility();
     const { hasMusdBalanceOnAnyChain } = useMusdBalance();
-    const hasSeenConversionEducationScreen = useSelector(
-      selectMusdConversionEducationSeen,
-    );
-    const isMoneyHomeEnabled = useSelector(selectMoneyHomeScreenEnabledFlag);
+    const { navigateToCash: handleViewCashTokens } = useCashNavigation();
 
     const isCashSectionEnabled = isMusdConversionEnabled && isGeoEligible;
-
-    const handleViewCashTokens = useCallback(() => {
-      if (isMoneyHomeEnabled) {
-        navigation.navigate(Routes.MONEY.ROOT);
-        return;
-      }
-
-      if (!hasSeenConversionEducationScreen) {
-        navigation.navigate(Routes.EARN.ROOT, {
-          screen: Routes.EARN.MUSD.CONVERSION_EDUCATION,
-          params: {
-            returnTo: { screen: Routes.WALLET.CASH_TOKENS_FULL_VIEW },
-          },
-        });
-        return;
-      }
-
-      navigation.navigate(Routes.WALLET.CASH_TOKENS_FULL_VIEW);
-    }, [isMoneyHomeEnabled, hasSeenConversionEducationScreen, navigation]);
 
     const { onLayout } = useHomeViewedEvent({
       sectionRef: sectionViewRef,
