+++ app/components/Views/Homepage/Sections/Cash/MusdAggregatedRow.tsx
@@ -31,10 +31,7 @@ import {
 } from '../../../../UI/Earn/constants/musd';
 import { useMusdBalance } from '../../../../UI/Earn/hooks/useMusdBalance';
 import { selectPrivacyMode } from '../../../../../selectors/preferencesController';
-import { selectMusdConversionEducationSeen } from '../../../../../reducers/user/selectors';
-import { useNavigation } from '@react-navigation/native';
-import Routes from '../../../../../constants/navigation/Routes';
-import { selectMoneyHomeScreenEnabledFlag } from '../../../../UI/Money/selectors/featureFlags';
+import { useCashNavigation } from './useCashNavigation';
 import { useMerklBonusClaim } from '../../../../UI/Earn/components/MerklRewards/hooks/useMerklBonusClaim';
 import { MUSD_EVENTS_CONSTANTS } from '../../../../UI/Earn/constants/events';
 import { LINEA_MUSD_ASSET_FOR_MERKL } from './CashGetMusdEmptyState.constants';
@@ -43,12 +40,7 @@ import { MetaMetricsEvents } from '../../../../../core/Analytics';
 
 const MusdAggregatedRow = () => {
   const tw = useTailwind();
-  const navigation = useNavigation();
   const privacyMode = useSelector(selectPrivacyMode);
-  const isMoneyHomeEnabled = useSelector(selectMoneyHomeScreenEnabledFlag);
-  const hasSeenConversionEducationScreen = useSelector(
-    selectMusdConversionEducationSeen,
-  );
   const { tokenBalanceAggregated, fiatBalanceAggregatedFormatted } =
     useMusdBalance();
   const { claimableReward, hasPendingClaim, isClaiming, claimRewards } =
@@ -76,22 +68,7 @@ const MusdAggregatedRow = () => {
     claimRewards();
   }, [trackEvent, createEventBuilder, claimRewards]);
 
-  const handleTokenRowPress = useCallback(() => {
-    if (isMoneyHomeEnabled) {
-      navigation.navigate(Routes.MONEY.ROOT);
-      return;
-    }
-    if (!hasSeenConversionEducationScreen) {
-      navigation.navigate(Routes.EARN.ROOT, {
-        screen: Routes.EARN.MUSD.CONVERSION_EDUCATION,
-        params: {
-          returnTo: { screen: Routes.WALLET.CASH_TOKENS_FULL_VIEW },
-        },
-      });
-      return;
-    }
-    navigation.navigate(Routes.WALLET.CASH_TOKENS_FULL_VIEW);
-  }, [navigation, isMoneyHomeEnabled, hasSeenConversionEducationScreen]);
+  const { navigateToCash: handleTokenRowPress } = useCashNavigation();
 
   const tokenBalanceDisplay = `${getIntlNumberFormatter(I18n.locale, {
     minimumFractionDigits: 0,
